VAB-600 Springboard BSP bare kernel modifications
The VAB-600 Springboard Board Support Package (BSP) supplies a
kernel based on Android/Linux 3.0.8 with some more patches
cherry-picked from later.

This commit intends to be a single changeset done to enable Springboard
support, over the likely starting point of the kernel development.
The rebase point is chosen by examining the number of differences
between the original and the BSP kernel, and chosing the one that
has the minimal changes I could find.